refactor(new-block-editor): reorganize component structure and update dialog services
