+++ core-web/libs/new-block-editor/src/lib/editor/components/floating-block-dialog.base.ts
