+++ core-web/libs/new-block-editor/src/lib/editor/components/image/image-dialog.service.ts
