+++ core-web/libs/new-block-editor/src/lib/editor/components/table/table-dialog.component.ts
