+++ core-web/libs/new-block-editor/src/lib/editor/components/video/video-dialog.service.ts
